Starting to split off the state of the data stream into its own class hierarchy.
