+++ scripts/build_lbann.sh
@@ -637,15 +637,15 @@ if [[ -n "${USER_MIRROR:-}" ]]; then
     MIRRORS="${MIRRORS:-} ${USER_MIRROR}"
 fi
 
-if [[ -n "${INSTALL_DEPS:-}" && -z "${SKIP_MIRRORS:-}" ]]; then
-    CMD="spack mirror add binary_mirror  https://binaries.spack.io/releases/v0.18"
-    echo ${CMD} | tee -a ${LOG}
-    [[ -z "${DRY_RUN:-}" ]] && { ${CMD} || exit_on_failure "${CMD}"; }
-    # Tell Spack to trust the keys in the build cache
-    CMD="spack buildcache keys --install --trust"
-    echo ${CMD} | tee -a ${LOG}
-    [[ -z "${DRY_RUN:-}" ]] && { ${CMD} || exit_on_failure "${CMD}"; }
-fi
+# if [[ -n "${INSTALL_DEPS:-}" && -z "${SKIP_MIRRORS:-}" ]]; then
+#     CMD="spack mirror add binary_mirror  https://binaries.spack.io/releases/v0.18"
+#     echo ${CMD} | tee -a ${LOG}
+#     [[ -z "${DRY_RUN:-}" ]] && { ${CMD} || exit_on_failure "${CMD}"; }
+#     # Tell Spack to trust the keys in the build cache
+#     CMD="spack buildcache keys --install --trust"
+#     echo ${CMD} | tee -a ${LOG}
+#     [[ -z "${DRY_RUN:-}" ]] && { ${CMD} || exit_on_failure "${CMD}"; }
+# fi
 
 if [[ -n "${INSTALL_DEPS:-}" && -n "${MIRRORS:-}" ]]; then
     i=0
